Correct a bug on file upload. Remove polymod plugin for users who has not rights to see them Correct a bug on email which cause emails to be marked as spam Add some little changes to polymod objects Change linx visibility during page edition Add missing translations on users and groups search Add bzrignore file Add little changes to 403 and 404 files

+++ automne/classes/files/filesManagement.php
@@ -953,7 +953,7 @@ class CMS_file extends CMS_grandFather
 	  * @static
 	  */
 	function getMaxUploadFileSize($unit = 'M') {
-		$max = (int) (ini_get("upload_max_filesize") < ini_get("post_max_size")) ? ini_get("upload_max_filesize") : ini_get("post_max_size");
+		$max = (int) ((int) ini_get("upload_max_filesize") < (int) ini_get("post_max_size")) ? ini_get("upload_max_filesize") : ini_get("post_max_size");
 		if ($unit == 'M') {
 			return io::substr($max, 0, -1);
 		} elseif($unit == 'K') {
